requirements.txt
# Main project dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
supabase==2.3.0
pydantic==2.5.0
python-dotenv==1.0.0